SO-4640: move common CodeSystem and Version related REST API calls...
...to test.commons
