+++ fhir/com.b2international.snowowl.fhir.rest.tests/.launch/fhir-unit-tests.launch
@@ -79,8 +79,8 @@
         <setEntry value="javax.servlet@default:default"/>
         <setEntry value="jaxb-api@default:default"/>
         <setEntry value="log4j.over.slf4j@default:default"/>
-        <setEntry value="net.bytebuddy.byte-buddy*1.9.0.v20181107-1410@default:default"/>
         <setEntry value="net.bytebuddy.byte-buddy-agent@default:default"/>
+        <setEntry value="net.bytebuddy.byte-buddy@default:default"/>
         <setEntry value="net.jodah.failsafe@default:default"/>
         <setEntry value="net.jodah.typetools@default:default"/>
         <setEntry value="org.antlr.runtime@default:default"/>
