+++ fhir/com.b2international.snowowl.fhir.rest.tests/src/com/b2international/snowowl/fhir/tests/TestArtifactCreator.java
@@ -15,14 +15,11 @@
  */
 package com.b2international.snowowl.fhir.tests;
 
-import java.util.Date;
-
 import com.b2international.snowowl.core.codesystem.CodeSystemRequests;
-import com.b2international.snowowl.core.date.DateFormats;
-import com.b2international.snowowl.core.date.EffectiveTimes;
 import com.b2international.snowowl.core.jobs.JobRequests;
 import com.b2international.snowowl.eventbus.IEventBus;
 import com.b2international.snowowl.test.commons.Services;
+import com.b2international.snowowl.test.commons.codesystem.CodeSystemVersionRestRequests;
 
 /**
  * Common superclass for test artefact creators
@@ -32,12 +29,12 @@ import com.b2international.snowowl.test.commons.Services;
 public class TestArtifactCreator {
 	
 	protected synchronized static void createVersion(String version, String codeSystemName) {
-		
+		String nextAvailableEffectiveTime = CodeSystemVersionRestRequests.getNextAvailableEffectiveDateAsString(codeSystemName);
 		String jobId = CodeSystemRequests.prepareNewCodeSystemVersion()
 			.setCodeSystemShortName(codeSystemName)
 			.setDescription("FHIR Test version")
 			.setVersionId(version)
-			.setEffectiveTime(EffectiveTimes.format(new Date(), DateFormats.SHORT))
+			.setEffectiveTime(nextAvailableEffectiveTime)
 			.buildAsync()
 			.runAsJob(String.format("Creating version '%s/%s'", codeSystemName, version))
 			.execute(getEventBus())
