+++ net4j/com.b2international.snowowl.eventbus.tests/.launch/eventbus-unit-tests.launch
@@ -39,9 +39,8 @@
         <setEntry value="ch.qos.logback.core@default:default"/>
         <setEntry value="com.diffplug.osgi.extension.sun.misc@default:false"/>
         <setEntry value="com.google.guava@default:default"/>
-        <setEntry value="net.bytebuddy.byte-buddy*1.9.0.v20181107-1410@default:default"/>
-        <setEntry value="net.bytebuddy.byte-buddy*1.9.3@default:default"/>
         <setEntry value="net.bytebuddy.byte-buddy-agent@default:default"/>
+        <setEntry value="net.bytebuddy.byte-buddy@default:default"/>
         <setEntry value="org.apache.felix.gogo.command@default:default"/>
         <setEntry value="org.apache.felix.gogo.runtime@default:default"/>
         <setEntry value="org.eclipse.core.contenttype@default:default"/>
