+++ tests/com.b2international.snowowl.test.commons/src/com/b2international/snowowl/test/commons/ApiTestConstants.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2021 B2i Healthcare Pte Ltd, http://b2i.sg
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.b2international.snowowl.test.commons;
+
+/**
+ * @since 7.15
+ */
+public interface ApiTestConstants {
+
+	/**
+	 * The context-relative base URL for the administrative controller. 
+	 */
+	String ADMIN_API = "/admin";
+	
+}
