+++ tests/com.b2international.snowowl.test.commons/src/com/b2international/snowowl/test/commons/codesystem/CodeSystemVersionRestRequests.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.b2international.snowowl.snomed.core.rest;
+package com.b2international.snowowl.test.commons.codesystem;
 
-import static com.b2international.snowowl.snomed.core.rest.CodeSystemRestRequests.createCodeSystem;
+import static com.b2international.snowowl.test.commons.codesystem.CodeSystemRestRequests.createCodeSystem;
 import static com.b2international.snowowl.test.commons.rest.RestExtensions.givenAuthenticatedRequest;
 
 import java.util.Calendar;
@@ -28,6 +28,7 @@ import java.util.TimeZone;
 import com.b2international.snowowl.core.api.IBranchPath;
 import com.b2international.snowowl.core.date.DateFormats;
 import com.b2international.snowowl.core.date.Dates;
+import com.b2international.snowowl.test.commons.ApiTestConstants;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSortedSet;
 import com.google.common.collect.Iterables;
@@ -48,7 +49,7 @@ public abstract class CodeSystemVersionRestRequests {
 	}
 	
 	public static ValidatableResponse getVersion(String shortName, String version) {
-		return givenAuthenticatedRequest(SnomedApiTestConstants.ADMIN_API)
+		return givenAuthenticatedRequest(ApiTestConstants.ADMIN_API)
 				.get("/codesystems/{shortName}/versions/{id}", shortName, version)
 				.then();
 	}
@@ -60,7 +61,7 @@ public abstract class CodeSystemVersionRestRequests {
 				.put("effectiveDate", effectiveDate)
 				.build();
 
-		return givenAuthenticatedRequest(SnomedApiTestConstants.ADMIN_API)
+		return givenAuthenticatedRequest(ApiTestConstants.ADMIN_API)
 				.contentType(ContentType.JSON)
 				.body(requestBody)
 				.post("/codesystems/{shortNameOrOid}/versions", shortName)
@@ -74,7 +75,7 @@ public abstract class CodeSystemVersionRestRequests {
 				.put("effectiveDate", effectiveDate)
 				.build();
 
-		return givenAuthenticatedRequest(SnomedApiTestConstants.ADMIN_API)
+		return givenAuthenticatedRequest(ApiTestConstants.ADMIN_API)
 				.contentType(ContentType.JSON)
 				.body(requestBody)
 				.post("/codesystems/{shortNameOrOid}/versions", shortName)
@@ -82,7 +83,7 @@ public abstract class CodeSystemVersionRestRequests {
 	}
 
 	public static ValidatableResponse getVersions(String shortName) {
-		return givenAuthenticatedRequest(SnomedApiTestConstants.ADMIN_API)
+		return givenAuthenticatedRequest(ApiTestConstants.ADMIN_API)
 				.and().contentType(ContentType.JSON)
 				.when()
 				.get("/codesystems/{shortName}/versions", shortName)
